requirements.txt
langchain==0.3.24
langchain_community==0.3.22
langchain_groq==0.3.2
langchain_huggingface==0.1.2
langchain_text_splitters==0.3.8
python-dotenv==1.1.0
streamlit==1.37.1
